+++ crates/bevy_pbr/Cargo.toml
@@ -29,4 +29,3 @@ bevy_window = { path = "../bevy_window", version = "0.5.0" }
 bitflags = "1.2"
 # direct dependency required for derive macro
 bytemuck = { version = "1", features = ["derive"] }
-wgpu = { version = "0.12.0", features = ["spirv"] }
