+++ crates/bevy_pbr/src/material.rs
@@ -1,321 +1,312 @@
-use crate::{AlphaMode, PbrPipeline, StandardMaterialFlags};
+use crate::{
+    AlphaMode, DrawMesh, MeshPipeline, MeshPipelineKey, MeshUniform, SetMeshBindGroup,
+    SetMeshViewBindGroup,
+};
 use bevy_app::{App, Plugin};
-use bevy_asset::{AddAsset, Handle, HandleUntyped};
-use bevy_ecs::system::{lifetimeless::SRes, SystemParamItem};
-use bevy_math::Vec4;
-use bevy_reflect::TypeUuid;
+use bevy_asset::{AddAsset, Asset, AssetServer, Handle};
+use bevy_core_pipeline::{AlphaMask3d, Opaque3d, Transparent3d};
+use bevy_ecs::{
+    entity::Entity,
+    prelude::World,
+    system::{
+        lifetimeless::{Read, SQuery, SRes},
+        Query, Res, ResMut, SystemParamItem,
+    },
+    world::FromWorld,
+};
 use bevy_render::{
-    color::Color,
-    render_asset::{PrepareAssetError, RenderAsset, RenderAssetPlugin, RenderAssets},
+    mesh::Mesh,
+    render_asset::{RenderAsset, RenderAssetPlugin, RenderAssets},
+    render_component::ExtractComponentPlugin,
+    render_phase::{
+        AddRenderCommand, DrawFunctions, EntityRenderCommand, RenderCommandResult, RenderPhase,
+        SetItemPipeline, TrackedRenderPass,
+    },
     render_resource::{
-        std140::{AsStd140, Std140},
-        BindGroup, Buffer, BufferInitDescriptor, BufferUsages,
+        BindGroup, BindGroupLayout, RenderPipelineCache, RenderPipelineDescriptor, Shader,
+        SpecializedPipeline, SpecializedPipelines,
     },
     renderer::RenderDevice,
-    texture::Image,
+    view::{ExtractedView, Msaa, VisibleEntities},
+    RenderApp, RenderStage,
 };
-use wgpu::{BindGroupDescriptor, BindGroupEntry, BindingResource};
+use std::hash::Hash;
+use std::marker::PhantomData;
 
-pub const DEFAULT_STANDARD_MATERIAL_HANDLE: HandleUntyped =
-    HandleUntyped::weak_from_u64(StandardMaterial::TYPE_UUID, 13142262394054731189);
-
-/// A material with "standard" properties used in PBR lighting
-/// Standard property values with pictures here
-/// <https://google.github.io/filament/Material%20Properties.pdf>.
-///
-/// May be created directly from a [`Color`] or an [`Image`].
-#[derive(Debug, Clone, TypeUuid)]
-#[uuid = "7494888b-c082-457b-aacf-517228cc0c22"]
-pub struct StandardMaterial {
-    /// Doubles as diffuse albedo for non-metallic, specular for metallic and a mix for everything
-    /// in between. If used together with a base_color_texture, this is factored into the final
-    /// base color as `base_color * base_color_texture_value`
-    pub base_color: Color,
-    pub base_color_texture: Option<Handle<Image>>,
-    // Use a color for user friendliness even though we technically don't use the alpha channel
-    // Might be used in the future for exposure correction in HDR
-    pub emissive: Color,
-    pub emissive_texture: Option<Handle<Image>>,
-    /// Linear perceptual roughness, clamped to [0.089, 1.0] in the shader
-    /// Defaults to minimum of 0.089
-    /// If used together with a roughness/metallic texture, this is factored into the final base
-    /// color as `roughness * roughness_texture_value`
-    pub perceptual_roughness: f32,
-    /// From [0.0, 1.0], dielectric to pure metallic
-    /// If used together with a roughness/metallic texture, this is factored into the final base
-    /// color as `metallic * metallic_texture_value`
-    pub metallic: f32,
-    pub metallic_roughness_texture: Option<Handle<Image>>,
-    /// Specular intensity for non-metals on a linear scale of [0.0, 1.0]
-    /// defaults to 0.5 which is mapped to 4% reflectance in the shader
-    pub reflectance: f32,
-    pub normal_map_texture: Option<Handle<Image>>,
-    pub occlusion_texture: Option<Handle<Image>>,
-    pub double_sided: bool,
-    pub unlit: bool,
-    pub alpha_mode: AlphaMode,
+pub trait Material: Asset + RenderAsset {
+    fn bind_group(render_asset: &<Self as RenderAsset>::PreparedAsset) -> &BindGroup;
+    fn bind_group_layout(render_device: &RenderDevice) -> BindGroupLayout;
+    #[allow(unused_variables)]
+    fn vertex_shader(asset_server: &AssetServer) -> Option<Handle<Shader>> {
+        None
+    }
+    #[allow(unused_variables)]
+    fn fragment_shader(asset_server: &AssetServer) -> Option<Handle<Shader>> {
+        None
+    }
+    #[allow(unused_variables)]
+    fn alpha_mode(render_asset: &<Self as RenderAsset>::PreparedAsset) -> AlphaMode {
+        AlphaMode::Opaque
+    }
 }
 
-impl Default for StandardMaterial {
+impl<M: Material> SpecializedMaterial for M {
+    type Key = ();
+
+    #[inline]
+    fn key(_render_asset: &<Self as RenderAsset>::PreparedAsset) -> Self::Key {}
+
+    #[inline]
+    fn specialize(_key: Self::Key, _descriptor: &mut RenderPipelineDescriptor) {}
+
+    #[inline]
+    fn bind_group(render_asset: &<Self as RenderAsset>::PreparedAsset) -> &BindGroup {
+        <M as Material>::bind_group(render_asset)
+    }
+
+    #[inline]
+    fn bind_group_layout(render_device: &RenderDevice) -> BindGroupLayout {
+        <M as Material>::bind_group_layout(render_device)
+    }
+
+    #[inline]
+    fn alpha_mode(render_asset: &<Self as RenderAsset>::PreparedAsset) -> AlphaMode {
+        <M as Material>::alpha_mode(render_asset)
+    }
+
+    #[inline]
+    fn vertex_shader(asset_server: &AssetServer) -> Option<Handle<Shader>> {
+        <M as Material>::vertex_shader(asset_server)
+    }
+
+    #[inline]
+    fn fragment_shader(asset_server: &AssetServer) -> Option<Handle<Shader>> {
+        <M as Material>::fragment_shader(asset_server)
+    }
+}
+
+pub trait SpecializedMaterial: Asset + RenderAsset {
+    type Key: PartialEq + Eq + Hash + Clone + Send + Sync;
+    fn key(render_asset: &<Self as RenderAsset>::PreparedAsset) -> Self::Key;
+    fn specialize(key: Self::Key, descriptor: &mut RenderPipelineDescriptor);
+    fn bind_group(render_asset: &<Self as RenderAsset>::PreparedAsset) -> &BindGroup;
+    fn bind_group_layout(render_device: &RenderDevice) -> BindGroupLayout;
+    #[allow(unused_variables)]
+    fn vertex_shader(asset_server: &AssetServer) -> Option<Handle<Shader>> {
+        None
+    }
+    #[allow(unused_variables)]
+    fn fragment_shader(asset_server: &AssetServer) -> Option<Handle<Shader>> {
+        None
+    }
+    #[allow(unused_variables)]
+    fn alpha_mode(render_asset: &<Self as RenderAsset>::PreparedAsset) -> AlphaMode {
+        AlphaMode::Opaque
+    }
+}
+
+pub struct MaterialPlugin<M: SpecializedMaterial>(PhantomData<M>);
+
+impl<M: SpecializedMaterial> Default for MaterialPlugin<M> {
     fn default() -> Self {
-        StandardMaterial {
-            base_color: Color::rgb(1.0, 1.0, 1.0),
-            base_color_texture: None,
-            emissive: Color::BLACK,
-            emissive_texture: None,
-            // This is the minimum the roughness is clamped to in shader code
-            // See <https://google.github.io/filament/Filament.html#materialsystem/parameterization/>
-            // It's the minimum floating point value that won't be rounded down to 0 in the
-            // calculations used. Although technically for 32-bit floats, 0.045 could be
-            // used.
-            perceptual_roughness: 0.089,
-            // Few materials are purely dielectric or metallic
-            // This is just a default for mostly-dielectric
-            metallic: 0.01,
-            metallic_roughness_texture: None,
-            // Minimum real-world reflectance is 2%, most materials between 2-5%
-            // Expressed in a linear scale and equivalent to 4% reflectance see
-            // <https://google.github.io/filament/Material%20Properties.pdf>
-            reflectance: 0.5,
-            occlusion_texture: None,
-            normal_map_texture: None,
-            double_sided: false,
-            unlit: false,
-            alpha_mode: AlphaMode::Opaque,
-        }
+        Self(Default::default())
     }
 }
 
-impl From<Color> for StandardMaterial {
-    fn from(color: Color) -> Self {
-        StandardMaterial {
-            base_color: color,
-            ..Default::default()
-        }
-    }
-}
-
-impl From<Handle<Image>> for StandardMaterial {
-    fn from(texture: Handle<Image>) -> Self {
-        StandardMaterial {
-            base_color_texture: Some(texture),
-            ..Default::default()
-        }
-    }
-}
-
-/// The GPU representation of the uniform data of a [`StandardMaterial`].
-#[derive(Clone, Default, AsStd140)]
-pub struct StandardMaterialUniformData {
-    /// Doubles as diffuse albedo for non-metallic, specular for metallic and a mix for everything
-    /// in between.
-    pub base_color: Vec4,
-    // Use a color for user friendliness even though we technically don't use the alpha channel
-    // Might be used in the future for exposure correction in HDR
-    pub emissive: Vec4,
-    /// Linear perceptual roughness, clamped to [0.089, 1.0] in the shader
-    /// Defaults to minimum of 0.089
-    pub roughness: f32,
-    /// From [0.0, 1.0], dielectric to pure metallic
-    pub metallic: f32,
-    /// Specular intensity for non-metals on a linear scale of [0.0, 1.0]
-    /// defaults to 0.5 which is mapped to 4% reflectance in the shader
-    pub reflectance: f32,
-    pub flags: u32,
-    /// When the alpha mode mask flag is set, any base color alpha above this cutoff means fully opaque,
-    /// and any below means fully transparent.
-    pub alpha_cutoff: f32,
-}
-
-/// This plugin adds the [`StandardMaterial`] asset to the app.
-pub struct StandardMaterialPlugin;
-
-impl Plugin for StandardMaterialPlugin {
+impl<M: SpecializedMaterial> Plugin for MaterialPlugin<M> {
     fn build(&self, app: &mut App) {
-        app.add_plugin(RenderAssetPlugin::<StandardMaterial>::default())
-            .add_asset::<StandardMaterial>();
+        app.add_asset::<M>()
+            .add_plugin(ExtractComponentPlugin::<Handle<M>>::default())
+            .add_plugin(RenderAssetPlugin::<M>::default());
+        app.sub_app(RenderApp)
+            .add_render_command::<Transparent3d, DrawMaterial<M>>()
+            .add_render_command::<Opaque3d, DrawMaterial<M>>()
+            .add_render_command::<AlphaMask3d, DrawMaterial<M>>()
+            .init_resource::<MaterialPipeline<M>>()
+            .init_resource::<SpecializedPipelines<MaterialPipeline<M>>>()
+            .add_system_to_stage(RenderStage::Queue, queue_material_meshes::<M>);
     }
 }
 
-/// The GPU representation of a [`StandardMaterial`].
-#[derive(Debug, Clone)]
-pub struct GpuStandardMaterial {
-    /// A buffer containing the [`StandardMaterialUniformData`] of the material.
-    pub buffer: Buffer,
-    /// The bind group specifying how the [`StandardMaterialUniformData`] and
-    /// all the textures of the material are bound.
-    pub bind_group: BindGroup,
-    pub has_normal_map: bool,
-    pub flags: StandardMaterialFlags,
-    pub base_color_texture: Option<Handle<Image>>,
-    pub alpha_mode: AlphaMode,
+pub struct MaterialPipeline<M: SpecializedMaterial> {
+    pub mesh_pipeline: MeshPipeline,
+    pub material_layout: BindGroupLayout,
+    pub vertex_shader: Option<Handle<Shader>>,
+    pub fragment_shader: Option<Handle<Shader>>,
+    marker: PhantomData<M>,
 }
 
-impl RenderAsset for StandardMaterial {
-    type ExtractedAsset = StandardMaterial;
-    type PreparedAsset = GpuStandardMaterial;
-    type Param = (
-        SRes<RenderDevice>,
-        SRes<PbrPipeline>,
-        SRes<RenderAssets<Image>>,
-    );
+impl<M: SpecializedMaterial> SpecializedPipeline for MaterialPipeline<M> {
+    type Key = (MeshPipelineKey, M::Key);
 
-    fn extract_asset(&self) -> Self::ExtractedAsset {
-        self.clone()
+    fn specialize(&self, key: Self::Key) -> RenderPipelineDescriptor {
+        let mut descriptor = self.mesh_pipeline.specialize(key.0);
+        if let Some(vertex_shader) = &self.vertex_shader {
+            descriptor.vertex.shader = vertex_shader.clone();
+        }
+
+        if let Some(fragment_shader) = &self.fragment_shader {
+            descriptor.fragment.as_mut().unwrap().shader = fragment_shader.clone();
+        }
+        descriptor.layout = Some(vec![
+            self.mesh_pipeline.view_layout.clone(),
+            self.material_layout.clone(),
+            self.mesh_pipeline.mesh_layout.clone(),
+        ]);
+
+        M::specialize(key.1, &mut descriptor);
+        descriptor
     }
+}
 
-    fn prepare_asset(
-        material: Self::ExtractedAsset,
-        (render_device, pbr_pipeline, gpu_images): &mut SystemParamItem<Self::Param>,
-    ) -> Result<Self::PreparedAsset, PrepareAssetError<Self::ExtractedAsset>> {
-        let (base_color_texture_view, base_color_sampler) = if let Some(result) = pbr_pipeline
-            .mesh_pipeline
-            .get_image_texture(gpu_images, &material.base_color_texture)
-        {
-            result
-        } else {
-            return Err(PrepareAssetError::RetryNextUpdate(material));
-        };
+impl<M: SpecializedMaterial> FromWorld for MaterialPipeline<M> {
+    fn from_world(world: &mut World) -> Self {
+        let asset_server = world.get_resource::<AssetServer>().unwrap();
+        let render_device = world.get_resource::<RenderDevice>().unwrap();
+        let material_layout = M::bind_group_layout(render_device);
 
-        let (emissive_texture_view, emissive_sampler) = if let Some(result) = pbr_pipeline
-            .mesh_pipeline
-            .get_image_texture(gpu_images, &material.emissive_texture)
-        {
-            result
-        } else {
-            return Err(PrepareAssetError::RetryNextUpdate(material));
-        };
+        MaterialPipeline {
+            mesh_pipeline: world.get_resource::<MeshPipeline>().unwrap().clone(),
+            material_layout,
+            vertex_shader: M::vertex_shader(asset_server),
+            fragment_shader: M::fragment_shader(asset_server),
+            marker: PhantomData,
+        }
+    }
+}
 
-        let (metallic_roughness_texture_view, metallic_roughness_sampler) = if let Some(result) =
-            pbr_pipeline
-                .mesh_pipeline
-                .get_image_texture(gpu_images, &material.metallic_roughness_texture)
-        {
-            result
-        } else {
-            return Err(PrepareAssetError::RetryNextUpdate(material));
-        };
-        let (normal_map_texture_view, normal_map_sampler) = if let Some(result) = pbr_pipeline
-            .mesh_pipeline
-            .get_image_texture(gpu_images, &material.normal_map_texture)
-        {
-            result
-        } else {
-            return Err(PrepareAssetError::RetryNextUpdate(material));
-        };
-        let (occlusion_texture_view, occlusion_sampler) = if let Some(result) = pbr_pipeline
-            .mesh_pipeline
-            .get_image_texture(gpu_images, &material.occlusion_texture)
-        {
-            result
-        } else {
-            return Err(PrepareAssetError::RetryNextUpdate(material));
-        };
-        let mut flags = StandardMaterialFlags::NONE;
-        if material.base_color_texture.is_some() {
-            flags |= StandardMaterialFlags::BASE_COLOR_TEXTURE;
-        }
-        if material.emissive_texture.is_some() {
-            flags |= StandardMaterialFlags::EMISSIVE_TEXTURE;
-        }
-        if material.metallic_roughness_texture.is_some() {
-            flags |= StandardMaterialFlags::METALLIC_ROUGHNESS_TEXTURE;
-        }
-        if material.occlusion_texture.is_some() {
-            flags |= StandardMaterialFlags::OCCLUSION_TEXTURE;
-        }
-        if material.double_sided {
-            flags |= StandardMaterialFlags::DOUBLE_SIDED;
-        }
-        if material.unlit {
-            flags |= StandardMaterialFlags::UNLIT;
-        }
-        let has_normal_map = material.normal_map_texture.is_some();
-        // NOTE: 0.5 is from the glTF default - do we want this?
-        let mut alpha_cutoff = 0.5;
-        match material.alpha_mode {
-            AlphaMode::Opaque => flags |= StandardMaterialFlags::ALPHA_MODE_OPAQUE,
-            AlphaMode::Mask(c) => {
-                alpha_cutoff = c;
-                flags |= StandardMaterialFlags::ALPHA_MODE_MASK
+type DrawMaterial<M> = (
+    SetItemPipeline,
+    SetMeshViewBindGroup<0>,
+    SetMaterialBindGroup<M, 1>,
+    SetMeshBindGroup<2>,
+    DrawMesh,
+);
+
+pub struct SetMaterialBindGroup<M: SpecializedMaterial, const I: usize>(PhantomData<M>);
+impl<M: SpecializedMaterial, const I: usize> EntityRenderCommand for SetMaterialBindGroup<M, I> {
+    type Param = (SRes<RenderAssets<M>>, SQuery<Read<Handle<M>>>);
+    fn render<'w>(
+        _view: Entity,
+        item: Entity,
+        (materials, query): SystemParamItem<'w, '_, Self::Param>,
+        pass: &mut TrackedRenderPass<'w>,
+    ) -> RenderCommandResult {
+        let material_handle = query.get(item).unwrap();
+        let material = materials.into_inner().get(material_handle).unwrap();
+        pass.set_bind_group(I, M::bind_group(material), &[]);
+        RenderCommandResult::Success
+    }
+}
+
+#[allow(clippy::too_many_arguments)]
+pub fn queue_material_meshes<M: SpecializedMaterial>(
+    opaque_draw_functions: Res<DrawFunctions<Opaque3d>>,
+    alpha_mask_draw_functions: Res<DrawFunctions<AlphaMask3d>>,
+    transparent_draw_functions: Res<DrawFunctions<Transparent3d>>,
+    material_pipeline: Res<MaterialPipeline<M>>,
+    mut pipelines: ResMut<SpecializedPipelines<MaterialPipeline<M>>>,
+    mut pipeline_cache: ResMut<RenderPipelineCache>,
+    msaa: Res<Msaa>,
+    render_meshes: Res<RenderAssets<Mesh>>,
+    render_materials: Res<RenderAssets<M>>,
+    material_meshes: Query<(&Handle<M>, &Handle<Mesh>, &MeshUniform)>,
+    mut views: Query<(
+        &ExtractedView,
+        &VisibleEntities,
+        &mut RenderPhase<Opaque3d>,
+        &mut RenderPhase<AlphaMask3d>,
+        &mut RenderPhase<Transparent3d>,
+    )>,
+) {
+    for (view, visible_entities, mut opaque_phase, mut alpha_mask_phase, mut transparent_phase) in
+        views.iter_mut()
+    {
+        let draw_opaque_pbr = opaque_draw_functions
+            .read()
+            .get_id::<DrawMaterial<M>>()
+            .unwrap();
+        let draw_alpha_mask_pbr = alpha_mask_draw_functions
+            .read()
+            .get_id::<DrawMaterial<M>>()
+            .unwrap();
+        let draw_transparent_pbr = transparent_draw_functions
+            .read()
+            .get_id::<DrawMaterial<M>>()
+            .unwrap();
+
+        let inverse_view_matrix = view.transform.compute_matrix().inverse();
+        let inverse_view_row_2 = inverse_view_matrix.row(2);
+        let mesh_key = MeshPipelineKey::from_msaa_samples(msaa.samples);
+
+        for visible_entity in &visible_entities.entities {
+            if let Ok((material_handle, mesh_handle, mesh_uniform)) =
+                material_meshes.get(*visible_entity)
+            {
+                if let Some(material) = render_materials.get(material_handle) {
+                    let mut mesh_key = mesh_key;
+                    if let Some(mesh) = render_meshes.get(mesh_handle) {
+                        if mesh.has_tangents {
+                            mesh_key |= MeshPipelineKey::VERTEX_TANGENTS;
+                        }
+                        mesh_key |=
+                            MeshPipelineKey::from_primitive_topology(mesh.primitive_topology);
+                    }
+                    let alpha_mode = M::alpha_mode(material);
+                    if let AlphaMode::Blend = alpha_mode {
+                        mesh_key |= MeshPipelineKey::TRANSPARENT_MAIN_PASS
+                    }
+
+                    let specialized_key = M::key(material);
+                    let pipeline_id = pipelines.specialize(
+                        &mut pipeline_cache,
+                        &material_pipeline,
+                        (mesh_key, specialized_key),
+                    );
+
+                    // NOTE: row 2 of the inverse view matrix dotted with column 3 of the model matrix
+                    // gives the z component of translation of the mesh in view space
+                    let mesh_z = inverse_view_row_2.dot(mesh_uniform.transform.col(3));
+                    match alpha_mode {
+                        AlphaMode::Opaque => {
+                            opaque_phase.add(Opaque3d {
+                                entity: *visible_entity,
+                                draw_function: draw_opaque_pbr,
+                                pipeline: pipeline_id,
+                                // NOTE: Front-to-back ordering for opaque with ascending sort means near should have the
+                                // lowest sort key and getting further away should increase. As we have
+                                // -z in front of the camera, values in view space decrease away from the
+                                // camera. Flipping the sign of mesh_z results in the correct front-to-back ordering
+                                distance: -mesh_z,
+                            });
+                        }
+                        AlphaMode::Mask(_) => {
+                            alpha_mask_phase.add(AlphaMask3d {
+                                entity: *visible_entity,
+                                draw_function: draw_alpha_mask_pbr,
+                                pipeline: pipeline_id,
+                                // NOTE: Front-to-back ordering for alpha mask with ascending sort means near should have the
+                                // lowest sort key and getting further away should increase. As we have
+                                // -z in front of the camera, values in view space decrease away from the
+                                // camera. Flipping the sign of mesh_z results in the correct front-to-back ordering
+                                distance: -mesh_z,
+                            });
+                        }
+                        AlphaMode::Blend => {
+                            transparent_phase.add(Transparent3d {
+                                entity: *visible_entity,
+                                draw_function: draw_transparent_pbr,
+                                pipeline: pipeline_id,
+                                // NOTE: Back-to-front ordering for transparent with ascending sort means far should have the
+                                // lowest sort key and getting closer should increase. As we have
+                                // -z in front of the camera, the largest distance is -far with values increasing toward the
+                                // camera. As such we can just use mesh_z as the distance
+                                distance: mesh_z,
+                            });
+                        }
+                    }
+                }
             }
-            AlphaMode::Blend => flags |= StandardMaterialFlags::ALPHA_MODE_BLEND,
-        };
-
-        let value = StandardMaterialUniformData {
-            base_color: material.base_color.as_linear_rgba_f32().into(),
-            emissive: material.emissive.into(),
-            roughness: material.perceptual_roughness,
-            metallic: material.metallic,
-            reflectance: material.reflectance,
-            flags: flags.bits(),
-            alpha_cutoff,
-        };
-        let value_std140 = value.as_std140();
-
-        let buffer = render_device.create_buffer_with_data(&BufferInitDescriptor {
-            label: Some("pbr_standard_material_uniform_buffer"),
-            usage: BufferUsages::UNIFORM | BufferUsages::COPY_DST,
-            contents: value_std140.as_bytes(),
-        });
-        let bind_group = render_device.create_bind_group(&BindGroupDescriptor {
-            entries: &[
-                BindGroupEntry {
-                    binding: 0,
-                    resource: buffer.as_entire_binding(),
-                },
-                BindGroupEntry {
-                    binding: 1,
-                    resource: BindingResource::TextureView(base_color_texture_view),
-                },
-                BindGroupEntry {
-                    binding: 2,
-                    resource: BindingResource::Sampler(base_color_sampler),
-                },
-                BindGroupEntry {
-                    binding: 3,
-                    resource: BindingResource::TextureView(emissive_texture_view),
-                },
-                BindGroupEntry {
-                    binding: 4,
-                    resource: BindingResource::Sampler(emissive_sampler),
-                },
-                BindGroupEntry {
-                    binding: 5,
-                    resource: BindingResource::TextureView(metallic_roughness_texture_view),
-                },
-                BindGroupEntry {
-                    binding: 6,
-                    resource: BindingResource::Sampler(metallic_roughness_sampler),
-                },
-                BindGroupEntry {
-                    binding: 7,
-                    resource: BindingResource::TextureView(occlusion_texture_view),
-                },
-                BindGroupEntry {
-                    binding: 8,
-                    resource: BindingResource::Sampler(occlusion_sampler),
-                },
-                BindGroupEntry {
-                    binding: 9,
-                    resource: BindingResource::TextureView(normal_map_texture_view),
-                },
-                BindGroupEntry {
-                    binding: 10,
-                    resource: BindingResource::Sampler(normal_map_sampler),
-                },
-            ],
-            label: Some("pbr_standard_material_bind_group"),
-            layout: &pbr_pipeline.material_layout,
-        });
-
-        Ok(GpuStandardMaterial {
-            buffer,
-            bind_group,
-            flags,
-            has_normal_map,
-            base_color_texture: material.base_color_texture,
-            alpha_mode: material.alpha_mode,
-        })
+        }
     }
 }
