+++ crates/bevy_pbr/src/render/mod.rs
@@ -3,341 +3,3 @@ mod mesh;
 
 pub use light::*;
 pub use mesh::*;
-use wgpu::SamplerBindingType;
-
-use crate::{AlphaMode, StandardMaterial, StandardMaterialUniformData, PBR_SHADER_HANDLE};
-use bevy_asset::Handle;
-use bevy_core_pipeline::{AlphaMask3d, Opaque3d, Transparent3d};
-use bevy_ecs::{
-    prelude::*,
-    system::{lifetimeless::*, SystemParamItem},
-};
-use bevy_render::{
-    mesh::Mesh,
-    render_asset::RenderAssets,
-    render_phase::{
-        DrawFunctions, EntityRenderCommand, RenderCommandResult, RenderPhase, SetItemPipeline,
-        TrackedRenderPass,
-    },
-    render_resource::{std140::AsStd140, *},
-    renderer::RenderDevice,
-    view::{ExtractedView, Msaa, VisibleEntities},
-};
-
-// NOTE: These must match the bit flags in bevy_pbr2/src/render/pbr.wgsl!
-bitflags::bitflags! {
-    #[repr(transparent)]
-    pub struct StandardMaterialFlags: u32 {
-        const BASE_COLOR_TEXTURE         = (1 << 0);
-        const EMISSIVE_TEXTURE           = (1 << 1);
-        const METALLIC_ROUGHNESS_TEXTURE = (1 << 2);
-        const OCCLUSION_TEXTURE          = (1 << 3);
-        const DOUBLE_SIDED               = (1 << 4);
-        const UNLIT                      = (1 << 5);
-        const ALPHA_MODE_OPAQUE          = (1 << 6);
-        const ALPHA_MODE_MASK            = (1 << 7);
-        const ALPHA_MODE_BLEND           = (1 << 8);
-        const NONE                       = 0;
-        const UNINITIALIZED              = 0xFFFF;
-    }
-}
-
-#[derive(Clone)]
-pub struct PbrPipeline {
-    pub mesh_pipeline: MeshPipeline,
-    pub material_layout: BindGroupLayout,
-}
-
-impl FromWorld for PbrPipeline {
-    fn from_world(world: &mut World) -> Self {
-        let render_device = world.get_resource::<RenderDevice>().unwrap();
-
-        let material_layout = render_device.create_bind_group_layout(&BindGroupLayoutDescriptor {
-            entries: &[
-                BindGroupLayoutEntry {
-                    binding: 0,
-                    visibility: ShaderStages::FRAGMENT,
-                    ty: BindingType::Buffer {
-                        ty: BufferBindingType::Uniform,
-                        has_dynamic_offset: false,
-                        min_binding_size: BufferSize::new(
-                            StandardMaterialUniformData::std140_size_static() as u64,
-                        ),
-                    },
-                    count: None,
-                },
-                // Base Color Texture
-                BindGroupLayoutEntry {
-                    binding: 1,
-                    visibility: ShaderStages::FRAGMENT,
-                    ty: BindingType::Texture {
-                        multisampled: false,
-                        sample_type: TextureSampleType::Float { filterable: true },
-                        view_dimension: TextureViewDimension::D2,
-                    },
-                    count: None,
-                },
-                // Base Color Texture Sampler
-                BindGroupLayoutEntry {
-                    binding: 2,
-                    visibility: ShaderStages::FRAGMENT,
-                    ty: BindingType::Sampler(SamplerBindingType::Filtering),
-                    count: None,
-                },
-                // Emissive Texture
-                BindGroupLayoutEntry {
-                    binding: 3,
-                    visibility: ShaderStages::FRAGMENT,
-                    ty: BindingType::Texture {
-                        multisampled: false,
-                        sample_type: TextureSampleType::Float { filterable: true },
-                        view_dimension: TextureViewDimension::D2,
-                    },
-                    count: None,
-                },
-                // Emissive Texture Sampler
-                BindGroupLayoutEntry {
-                    binding: 4,
-                    visibility: ShaderStages::FRAGMENT,
-                    ty: BindingType::Sampler(SamplerBindingType::Filtering),
-                    count: None,
-                },
-                // Metallic Roughness Texture
-                BindGroupLayoutEntry {
-                    binding: 5,
-                    visibility: ShaderStages::FRAGMENT,
-                    ty: BindingType::Texture {
-                        multisampled: false,
-                        sample_type: TextureSampleType::Float { filterable: true },
-                        view_dimension: TextureViewDimension::D2,
-                    },
-                    count: None,
-                },
-                // Metallic Roughness Texture Sampler
-                BindGroupLayoutEntry {
-                    binding: 6,
-                    visibility: ShaderStages::FRAGMENT,
-                    ty: BindingType::Sampler(SamplerBindingType::Filtering),
-                    count: None,
-                },
-                // Occlusion Texture
-                BindGroupLayoutEntry {
-                    binding: 7,
-                    visibility: ShaderStages::FRAGMENT,
-                    ty: BindingType::Texture {
-                        multisampled: false,
-                        sample_type: TextureSampleType::Float { filterable: true },
-                        view_dimension: TextureViewDimension::D2,
-                    },
-                    count: None,
-                },
-                // Occlusion Texture Sampler
-                BindGroupLayoutEntry {
-                    binding: 8,
-                    visibility: ShaderStages::FRAGMENT,
-                    ty: BindingType::Sampler(SamplerBindingType::Filtering),
-                    count: None,
-                },
-                // Normal Map Texture
-                BindGroupLayoutEntry {
-                    binding: 9,
-                    visibility: ShaderStages::FRAGMENT,
-                    ty: BindingType::Texture {
-                        multisampled: false,
-                        sample_type: TextureSampleType::Float { filterable: true },
-                        view_dimension: TextureViewDimension::D2,
-                    },
-                    count: None,
-                },
-                // Normal Map Texture Sampler
-                BindGroupLayoutEntry {
-                    binding: 10,
-                    visibility: ShaderStages::FRAGMENT,
-                    ty: BindingType::Sampler(SamplerBindingType::Filtering),
-                    count: None,
-                },
-            ],
-            label: Some("pbr_material_layout"),
-        });
-
-        PbrPipeline {
-            material_layout,
-            mesh_pipeline: world.get_resource::<MeshPipeline>().unwrap().clone(),
-        }
-    }
-}
-
-#[derive(Clone, Copy, Hash, PartialEq, Eq)]
-pub struct PbrPipelineKey {
-    pub mesh_key: MeshPipelineKey,
-    pub normal_map: bool,
-}
-
-impl SpecializedPipeline for PbrPipeline {
-    type Key = PbrPipelineKey;
-
-    fn specialize(&self, key: Self::Key) -> RenderPipelineDescriptor {
-        let mut descriptor = self.mesh_pipeline.specialize(key.mesh_key);
-        descriptor.fragment.as_mut().unwrap().shader = PBR_SHADER_HANDLE.typed::<Shader>();
-        descriptor.layout = Some(vec![
-            self.mesh_pipeline.view_layout.clone(),
-            self.material_layout.clone(),
-            self.mesh_pipeline.mesh_layout.clone(),
-        ]);
-
-        if key.normal_map {
-            descriptor
-                .fragment
-                .as_mut()
-                .unwrap()
-                .shader_defs
-                .push(String::from("STANDARDMATERIAL_NORMAL_MAP"));
-        }
-        if let Some(label) = &mut descriptor.label {
-            *label = format!("pbr_{}", *label).into();
-        }
-        descriptor
-    }
-}
-
-pub struct PbrViewBindGroup {
-    pub value: BindGroup,
-}
-
-#[allow(clippy::too_many_arguments)]
-pub fn queue_meshes(
-    opaque_draw_functions: Res<DrawFunctions<Opaque3d>>,
-    alpha_mask_draw_functions: Res<DrawFunctions<AlphaMask3d>>,
-    transparent_draw_functions: Res<DrawFunctions<Transparent3d>>,
-    pbr_pipeline: Res<PbrPipeline>,
-    mut pipelines: ResMut<SpecializedPipelines<PbrPipeline>>,
-    mut pipeline_cache: ResMut<RenderPipelineCache>,
-    msaa: Res<Msaa>,
-    render_meshes: Res<RenderAssets<Mesh>>,
-    render_materials: Res<RenderAssets<StandardMaterial>>,
-    standard_material_meshes: Query<(&Handle<StandardMaterial>, &Handle<Mesh>, &MeshUniform)>,
-    mut views: Query<(
-        &ExtractedView,
-        &VisibleEntities,
-        &mut RenderPhase<Opaque3d>,
-        &mut RenderPhase<AlphaMask3d>,
-        &mut RenderPhase<Transparent3d>,
-    )>,
-) {
-    for (view, visible_entities, mut opaque_phase, mut alpha_mask_phase, mut transparent_phase) in
-        views.iter_mut()
-    {
-        let draw_opaque_pbr = opaque_draw_functions.read().get_id::<DrawPbr>().unwrap();
-        let draw_alpha_mask_pbr = alpha_mask_draw_functions
-            .read()
-            .get_id::<DrawPbr>()
-            .unwrap();
-        let draw_transparent_pbr = transparent_draw_functions
-            .read()
-            .get_id::<DrawPbr>()
-            .unwrap();
-
-        let inverse_view_matrix = view.transform.compute_matrix().inverse();
-        let inverse_view_row_2 = inverse_view_matrix.row(2);
-        let mesh_key = MeshPipelineKey::from_msaa_samples(msaa.samples);
-
-        for visible_entity in &visible_entities.entities {
-            if let Ok((material_handle, mesh_handle, mesh_uniform)) =
-                standard_material_meshes.get(*visible_entity)
-            {
-                if let Some(material) = render_materials.get(material_handle) {
-                    let mut pbr_key = PbrPipelineKey {
-                        mesh_key,
-                        normal_map: material.has_normal_map,
-                    };
-                    if let Some(mesh) = render_meshes.get(mesh_handle) {
-                        if mesh.has_tangents {
-                            pbr_key.mesh_key |= MeshPipelineKey::VERTEX_TANGENTS;
-                        }
-                        pbr_key.mesh_key |=
-                            MeshPipelineKey::from_primitive_topology(mesh.primitive_topology);
-                    }
-
-                    if let AlphaMode::Blend = material.alpha_mode {
-                        pbr_key.mesh_key |= MeshPipelineKey::TRANSPARENT_MAIN_PASS
-                    }
-
-                    let pipeline_id =
-                        pipelines.specialize(&mut pipeline_cache, &pbr_pipeline, pbr_key);
-
-                    // NOTE: row 2 of the inverse view matrix dotted with column 3 of the model matrix
-                    // gives the z component of translation of the mesh in view space
-                    let mesh_z = inverse_view_row_2.dot(mesh_uniform.transform.col(3));
-                    match material.alpha_mode {
-                        AlphaMode::Opaque => {
-                            opaque_phase.add(Opaque3d {
-                                entity: *visible_entity,
-                                draw_function: draw_opaque_pbr,
-                                pipeline: pipeline_id,
-                                // NOTE: Front-to-back ordering for opaque with ascending sort means near should have the
-                                // lowest sort key and getting further away should increase. As we have
-                                // -z in front of the camera, values in view space decrease away from the
-                                // camera. Flipping the sign of mesh_z results in the correct front-to-back ordering
-                                distance: -mesh_z,
-                            });
-                        }
-                        AlphaMode::Mask(_) => {
-                            alpha_mask_phase.add(AlphaMask3d {
-                                entity: *visible_entity,
-                                draw_function: draw_alpha_mask_pbr,
-                                pipeline: pipeline_id,
-                                // NOTE: Front-to-back ordering for alpha mask with ascending sort means near should have the
-                                // lowest sort key and getting further away should increase. As we have
-                                // -z in front of the camera, values in view space decrease away from the
-                                // camera. Flipping the sign of mesh_z results in the correct front-to-back ordering
-                                distance: -mesh_z,
-                            });
-                        }
-                        AlphaMode::Blend => {
-                            transparent_phase.add(Transparent3d {
-                                entity: *visible_entity,
-                                draw_function: draw_transparent_pbr,
-                                pipeline: pipeline_id,
-                                // NOTE: Back-to-front ordering for transparent with ascending sort means far should have the
-                                // lowest sort key and getting closer should increase. As we have
-                                // -z in front of the camera, the largest distance is -far with values increasing toward the
-                                // camera. As such we can just use mesh_z as the distance
-                                distance: mesh_z,
-                            });
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-pub type DrawPbr = (
-    SetItemPipeline,
-    SetMeshViewBindGroup<0>,
-    SetStandardMaterialBindGroup<1>,
-    SetMeshBindGroup<2>,
-    DrawMesh,
-);
-
-pub struct SetStandardMaterialBindGroup<const I: usize>;
-impl<const I: usize> EntityRenderCommand for SetStandardMaterialBindGroup<I> {
-    type Param = (
-        SRes<RenderAssets<StandardMaterial>>,
-        SQuery<Read<Handle<StandardMaterial>>>,
-    );
-    #[inline]
-    fn render<'w>(
-        _view: Entity,
-        item: Entity,
-        (materials, handle_query): SystemParamItem<'w, '_, Self::Param>,
-        pass: &mut TrackedRenderPass<'w>,
-    ) -> RenderCommandResult {
-        let handle = handle_query.get(item).unwrap();
-        let materials = materials.into_inner();
-        let material = materials.get(handle).unwrap();
-
-        pass.set_bind_group(I, &material.bind_group, &[]);
-        RenderCommandResult::Success
-    }
-}
