+++ examples/shader/shader_material.rs
@@ -1,34 +1,22 @@
 use bevy::{
-    core_pipeline::Transparent3d,
-    ecs::system::{lifetimeless::*, SystemParamItem},
-    pbr::{
-        DrawMesh, MeshPipeline, MeshPipelineKey, MeshUniform, SetMeshBindGroup,
-        SetMeshViewBindGroup,
-    },
+    ecs::system::{lifetimeless::SRes, SystemParamItem},
+    pbr::MaterialPipeline,
     prelude::*,
     reflect::TypeUuid,
     render::{
-        camera::PerspectiveCameraBundle,
-        render_asset::{PrepareAssetError, RenderAsset, RenderAssetPlugin, RenderAssets},
-        render_component::ExtractComponentPlugin,
-        render_phase::{
-            AddRenderCommand, DrawFunctions, EntityRenderCommand, RenderCommandResult, RenderPhase,
-            SetItemPipeline, TrackedRenderPass,
-        },
+        render_asset::{PrepareAssetError, RenderAsset},
         render_resource::{
             std140::{AsStd140, Std140},
             *,
         },
         renderer::RenderDevice,
-        view::{ExtractedView, Msaa},
-        RenderApp, RenderStage,
     },
 };
 
 fn main() {
     App::new()
         .add_plugins(DefaultPlugins)
-        .add_plugin(CustomMaterialPlugin)
+        .add_plugin(MaterialPlugin::<CustomMaterial>::default())
         .add_startup_system(setup)
         .run();
 }
@@ -40,16 +28,14 @@ fn setup(
     mut materials: ResMut<Assets<CustomMaterial>>,
 ) {
     // cube
-    commands.spawn().insert_bundle((
-        meshes.add(Mesh::from(shape::Cube { size: 1.0 })),
-        Transform::from_xyz(0.0, 0.5, 0.0),
-        GlobalTransform::default(),
-        Visibility::default(),
-        ComputedVisibility::default(),
-        materials.add(CustomMaterial {
+    commands.spawn().insert_bundle(MaterialMeshBundle {
+        mesh: meshes.add(Mesh::from(shape::Cube { size: 1.0 })),
+        transform: Transform::from_xyz(0.0, 0.5, 0.0),
+        material: materials.add(CustomMaterial {
             color: Color::GREEN,
         }),
-    ));
+        ..Default::default()
+    });
 
     // camera
     commands.spawn_bundle(PerspectiveCameraBundle {
@@ -73,14 +59,14 @@ pub struct GpuCustomMaterial {
 impl RenderAsset for CustomMaterial {
     type ExtractedAsset = CustomMaterial;
     type PreparedAsset = GpuCustomMaterial;
-    type Param = (SRes<RenderDevice>, SRes<CustomPipeline>);
+    type Param = (SRes<RenderDevice>, SRes<MaterialPipeline<Self>>);
     fn extract_asset(&self) -> Self::ExtractedAsset {
         self.clone()
     }
 
     fn prepare_asset(
         extracted_asset: Self::ExtractedAsset,
-        (render_device, custom_pipeline): &mut SystemParamItem<Self::Param>,
+        (render_device, material_pipeline): &mut SystemParamItem<Self::Param>,
     ) -> Result<Self::PreparedAsset, PrepareAssetError<Self::ExtractedAsset>> {
         let color = Vec4::from_slice(&extracted_asset.color.as_linear_rgba_f32());
         let buffer = render_device.create_buffer_with_data(&BufferInitDescriptor {
@@ -94,7 +80,7 @@ impl RenderAsset for CustomMaterial {
                 resource: buffer.as_entire_binding(),
             }],
             label: None,
-            layout: &custom_pipeline.material_layout,
+            layout: &material_pipeline.material_layout,
         });
 
         Ok(GpuCustomMaterial {
@@ -103,51 +89,18 @@ impl RenderAsset for CustomMaterial {
         })
     }
 }
-pub struct CustomMaterialPlugin;
 
-impl Plugin for CustomMaterialPlugin {
-    fn build(&self, app: &mut App) {
-        app.add_asset::<CustomMaterial>()
-            .add_plugin(ExtractComponentPlugin::<Handle<CustomMaterial>>::default())
-            .add_plugin(RenderAssetPlugin::<CustomMaterial>::default());
-        app.sub_app_mut(RenderApp)
-            .add_render_command::<Transparent3d, DrawCustom>()
-            .init_resource::<CustomPipeline>()
-            .init_resource::<SpecializedPipelines<CustomPipeline>>()
-            .add_system_to_stage(RenderStage::Queue, queue_custom);
+impl Material for CustomMaterial {
+    fn fragment_shader(asset_server: &AssetServer) -> Option<Handle<Shader>> {
+        Some(asset_server.load("shaders/custom_material.wgsl"))
     }
-}
 
-pub struct CustomPipeline {
-    mesh_pipeline: MeshPipeline,
-    material_layout: BindGroupLayout,
-    shader: Handle<Shader>,
-}
-
-impl SpecializedPipeline for CustomPipeline {
-    type Key = MeshPipelineKey;
-
-    fn specialize(&self, key: Self::Key) -> RenderPipelineDescriptor {
-        let mut descriptor = self.mesh_pipeline.specialize(key);
-        descriptor.fragment.as_mut().unwrap().shader = self.shader.clone();
-        descriptor.layout = Some(vec![
-            self.mesh_pipeline.view_layout.clone(),
-            self.material_layout.clone(),
-            self.mesh_pipeline.mesh_layout.clone(),
-        ]);
-        descriptor
+    fn bind_group(render_asset: &<Self as RenderAsset>::PreparedAsset) -> &BindGroup {
+        &render_asset.bind_group
     }
-}
 
-impl FromWorld for CustomPipeline {
-    fn from_world(world: &mut World) -> Self {
-        let asset_server = world.get_resource::<AssetServer>().unwrap();
-        // Watch for changes, allowing for hot shader reloading
-        // Try changing custom_material.wgsl while the app is running!
-        asset_server.watch_for_changes().unwrap();
-
-        let render_device = world.get_resource::<RenderDevice>().unwrap();
-        let material_layout = render_device.create_bind_group_layout(&BindGroupLayoutDescriptor {
+    fn bind_group_layout(render_device: &RenderDevice) -> BindGroupLayout {
+        render_device.create_bind_group_layout(&BindGroupLayoutDescriptor {
             entries: &[BindGroupLayoutEntry {
                 binding: 0,
                 visibility: ShaderStages::FRAGMENT,
@@ -159,80 +112,6 @@ impl FromWorld for CustomPipeline {
                 count: None,
             }],
             label: None,
-        });
-
-        CustomPipeline {
-            mesh_pipeline: world.get_resource::<MeshPipeline>().unwrap().clone(),
-            shader: asset_server.load("shaders/custom_material.wgsl"),
-            material_layout,
-        }
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-pub fn queue_custom(
-    transparent_3d_draw_functions: Res<DrawFunctions<Transparent3d>>,
-    materials: Res<RenderAssets<CustomMaterial>>,
-    render_meshes: Res<RenderAssets<Mesh>>,
-    custom_pipeline: Res<CustomPipeline>,
-    mut pipeline_cache: ResMut<RenderPipelineCache>,
-    mut specialized_pipelines: ResMut<SpecializedPipelines<CustomPipeline>>,
-    msaa: Res<Msaa>,
-    material_meshes: Query<(Entity, &Handle<CustomMaterial>, &Handle<Mesh>, &MeshUniform)>,
-    mut views: Query<(&ExtractedView, &mut RenderPhase<Transparent3d>)>,
-) {
-    let draw_custom = transparent_3d_draw_functions
-        .read()
-        .get_id::<DrawCustom>()
-        .unwrap();
-    let key = MeshPipelineKey::from_msaa_samples(msaa.samples);
-    for (view, mut transparent_phase) in views.iter_mut() {
-        let view_matrix = view.transform.compute_matrix();
-        let view_row_2 = view_matrix.row(2);
-        for (entity, material_handle, mesh_handle, mesh_uniform) in material_meshes.iter() {
-            if materials.contains_key(material_handle) {
-                if let Some(mesh) = render_meshes.get(mesh_handle) {
-                    let key =
-                        key | MeshPipelineKey::from_primitive_topology(mesh.primitive_topology);
-                    transparent_phase.add(Transparent3d {
-                        entity,
-                        pipeline: specialized_pipelines.specialize(
-                            &mut pipeline_cache,
-                            &custom_pipeline,
-                            key,
-                        ),
-                        draw_function: draw_custom,
-                        distance: view_row_2.dot(mesh_uniform.transform.col(3)),
-                    });
-                }
-            }
-        }
-    }
-}
-
-type DrawCustom = (
-    SetItemPipeline,
-    SetMeshViewBindGroup<0>,
-    SetCustomMaterialBindGroup,
-    SetMeshBindGroup<2>,
-    DrawMesh,
-);
-
-struct SetCustomMaterialBindGroup;
-impl EntityRenderCommand for SetCustomMaterialBindGroup {
-    type Param = (
-        SRes<RenderAssets<CustomMaterial>>,
-        SQuery<Read<Handle<CustomMaterial>>>,
-    );
-    fn render<'w>(
-        _view: Entity,
-        item: Entity,
-        (materials, query): SystemParamItem<'w, '_, Self::Param>,
-        pass: &mut TrackedRenderPass<'w>,
-    ) -> RenderCommandResult {
-        let material_handle = query.get(item).unwrap();
-        let material = materials.into_inner().get(material_handle).unwrap();
-        pass.set_bind_group(1, &material.bind_group, &[]);
-        RenderCommandResult::Success
+        })
     }
 }
